fix delete button

+++ frontend/src/app/pages/agenda/components/card-next-events/card-next-events.ts
@@ -1,7 +1,8 @@
-import { Component, inject, input } from '@angular/core';
+import { Component, inject, input, output } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FaIconButton } from '../../../../shared/buttons/fa-icon-button/fa-icon-button';
 import { EventService } from '../../../../core/services/event.service';
+import { PopupService } from '../../../../shared/popup-modal/popup-modal.service';
 
 @Component({
   selector: 'app-card-next-events',
@@ -11,6 +12,9 @@ import { EventService } from '../../../../core/services/event.service';
 export class CardNextEvents {
 
   private eventService = inject(EventService)
+  private popupService = inject(PopupService)
+
+  deleted = output<void>();
 
   id = input.required<string>();
   titulo = input.required<string>();
@@ -19,16 +23,18 @@ export class CardNextEvents {
   dataFim = input.required<string>();
   cor = input.required<string>();
 
+
   get formattedDate(): string {
     const start = new Date(this.dataInicio());
     const end = new Date(this.dataFim());
+
 
     const data = start.toLocaleDateString('pt-BR', {
       day: '2-digit',
       month: 'short',
     });
 
-    const datFim = end.toLocaleDateString('pt-BR',{
+    const datFim = end.toLocaleDateString('pt-BR', {
       day: '2-digit',
       month: 'short',
     });
@@ -45,11 +51,22 @@ export class CardNextEvents {
 
     return `${data} - ${startTime} / ${datFim} - ${endTime}`;
   }
-  
+
   onDelete() {
-  this.eventService.deleteEvent(this.id())
-    .subscribe({
-      error: (err) => console.error('Erro ao excluir evento', err)
-    });
-}
+    this.popupService.showConfirm(
+      'Tem certeza que deseja excluir este evento?',
+      () => {
+        this.eventService.deleteEvent(this.id())
+          .subscribe({
+            next: () => {
+              this.deleted.emit();
+              this.popupService.showSuccess('Evento excluído com sucesso!');
+            },
+            error: (err) => {
+              this.popupService.showError('Erro ao excluir evento.' + err);
+            }
+          });
+      }
+    )
+  }
 }
